Test (#1)
* Exclude overlays from the app counter

This exclude overlays from the app counter that shows
in manage app summary.

idea from @KreAch3R

Change-Id: I8cd591dc89eb87057e8c2ae3b2cca2c76c34534b
(cherry picked from commit 990eeb53387bbd52e76ce7e7b45c171d1e01e7a1)

* Hide the "show/hide overlays" when no overlay installed

Change-Id: I54088d27c9492e0992992de59bf340fe2465260a

* Expose color for external settings icons

Marhsmallow now allows for system apps to add settings.  An icon is
loaded from the external package and a tint is applied to the icon
so that it matches other icons in the dashboard.  The tint color
uses the colorAccent attribute from the currently applied style.
Themes can overlay the colorAccent but this color is also used for
other elements of the UI which may not be ideal.

This patch adds a new color external_tile_icon_tint_color which
points to the attribute ?android:attr/colorAccent when not themed.
This retains the original look but offers better flexibility for
theme designers.

- Ported up to 7.1.1

Change-Id: Ifb4a99a315271cdcf978a796e3dea0eaef8d7750

Conflicts:
	res/values/cm_colors.xml

* Settings: Set root access options appropriately

It is possible to be running a user build with a debuggable boot image.
In this case, "su" will not be available.  So only show none/adb.

Issue-Id: BACON-4461
Change-Id: Iaa7df8311b9ea81eabb1566ba6f9159fdc9fab34