Settings: Always show Privacy Guard permissions
Enabling Privacy Guard for an app simply means switching a set of
operations to MODE_ASK, independently on whether the application
actually declared those ops (though a permission) or not. The
framework keeps track only of the ops with a non-default value. As
consequence, all the ops set by Privacy Guard that aren't declared
by the app through its manifest are effectively lost when set to
their default value and the settings won't show them.

Never hide the Privacy Guard ops to provide a consistent UI.

Change-Id: Iafcf058f5e2074982bf45f8c82ef8d027b9358f0
Signed-off-by: DarkKnight6499 <yazad.madan@gmail.com>